Fix semantics of initial root element in DOM rendering
Diff and patch root element is not anymore a child of the
given container (to the DOM Driver). It is now the given
container itself. Diff and patch will not override and destroy
the identity of the given container. Instead, it will preserve it.
In fact, the given container element is parsed to a vtree, and is
used as the first vtree in the diff & patch sequence. This enables
avoiding to rerender the same tree client-side if it was already
rendered server-side, which saves processing power especially on
mobile devices. The semantics of the root element Observable,
get(':root'), for the DOM Driver has also changed: the first
emission is what was given as the container. This is important for
attaching event listeners.