perf(events): introduce EventDelegator and Map
DOMSource.events() was attaching a DOM event listener for each call to
events(). This was not a good thing since with N events() of the same
parameters, there would be N event listeners, and N bubbling
simulations. EventDelegator fixes that by having just one bubbling
simulation for all those N.