fix(DOMSource): attach DOM listeners also on ready root
To support a root container which comes ready-rendered from the server
(i.e. the client should NOT re-render the same through the virtual DOM),
we need to attach DOM event listeners also on the given root container.
We had to update the logic in setupEventDelegatorOnTopElement to also be
able to update the topElement inside the currently existing
EventDelegator, instead of reusing the old element which may be
irrelevant. Also, to avoid race conditions and obsolete data related to
MemoryStream and startWith, we need to make rootElement$ an infinite
stream (no complete). This also means every Stream<Event> returned by
DOMSource.events() is infinite, but the semantics here seem right: we
don't expect an explicit end to the interaction on the client-side.