* fix(isolation): keep EventDelegator up to date with isolated Element
* test(isolation): make sure eventlisteners are properly updated

Test that events are still handled correctly when isolated elements
are added and removed. Thank you @whitecolor for this test!

* fix(isolation): keep EventDelegator up to date with isolated Element

When attaching event listners to elements other than the root element
supplied to makeDOMDriver(), the elements are likely to be removed
and recreated. This fix makes sure that EventDelegators always have
the latest element created by snabbdom for an isolation boundary.

PR #121