fix(transposition): fix transpose bug of mutability
The diff and patch procedure of snabbdom keeps the old vnode reference,
but in xs.combine, the array of recent values is mutated in xstream's
implementation, leading to cases where the old vnode had the same
contents as the new vnode, which means diff procedure would see no
difference, and no patch would happen. This bug is fixed by cloning the
results of xs.combine.